requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
configparser==5.2.6
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==l.11.8
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
sex==1.16.0
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
zipp==3.8.0